Scan update
