+++ uscope/imagep/thread.py
@@ -3,35 +3,14 @@ from uscope.imager.imager_util import get_scaled
 from uscope.imagep.pipeline import CSImageProcessor
 from uscope.imager.autofocus import Autofocus
 from uscope.threads import CommandThreadBase
+from uscope.imagep.util import find_qr_code_match
 
 import threading
 import queue
 import traceback
 from PIL import Image
 from uscope.microscope import MicroscopeStop
-import re
 import os.path
-from pyzbar.pyzbar import decode as pyzbar_decode
-
-
-def find_qr_code_match(image, expr):
-    """
-    Scan image for QR code(s) and return first match
-    """
-    decode_results = pyzbar_decode(image)
-    for decoded in decode_results:
-        data = decoded.data
-        try:
-            m = re.match(expr, str(data, "utf-8"))
-            if m:
-                return m.string
-        except TypeError as e:
-            # Log this error?
-            print("Invalid qr_regex", e)
-            return None
-
-    # No detected match
-    return None
 
 
 class ImageProcessingThreadBase(CommandThreadBase):
